backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
firebase-admin==6.5.0
pydantic==2.8.2
python-dotenv==1.0.1
openai==1.12.0
